fix(issue-243): Fixing UI issue + refactoring of ChatResponsePanel
